fix: 버전 체크후 코어 데이터 초기화 로직 수정

+++ Tekken8 Frame Data/TK8/Utility/FrameDataVersionManager.swift
@@ -22,6 +22,11 @@ final class FrameDataVersionManager: FrameDataVersionManageable {
         let localVersion = fetchLocalVersion()
         let serverVesion = try await supabaseManager.fetchVersion()
         
+        if localVersion == 0 {
+            updateLocalVersion(version: serverVesion)
+            return
+        }
+        
         if localVersion < serverVesion {
             try coreDataManager.deleteAll()
             updateLocalVersion(version: serverVesion)
